# api_server/docs/swagger.yaml
basePath: /api
definitions:
  errors.APIError:
    properties:
      code:
        $ref: '#/definitions/errors.APIErrorCode'
      error:
        type: string
    type: object
  errors.APIErrorCode:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - ErrCodeMissingParam
    - ErrCodeIncorrectParam
    - ErrCodeSQLNoRows
    - ErrCodeSQLInternalError
  types.APIResponse:
    properties:
      error:
        $ref: '#/definitions/errors.APIError'
      message:
        type: string
    type: object
...
  types.Expense:
    properties:
      expense_amount:
        type: number
      expense_date:
        type: string
      expense_id:
        type: integer
      last_edited:
        type: string
    type: object
...
info:
  contact: {}
  title: HACS database API
  version: "1.0"
paths:
...
  /expense/all:
    get:
      description: Get all expenses
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Expense'
            type: array
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get all expenses
      tags:
      - expense
  /expense/date/{date}:
    get:
      description: Get expenses by expense_date
      parameters:
      - description: Expense date
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Expense'
            type: array
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get expenses by expense_date
      tags:
      - expense
  /expense/date/range:
    post:
      description: Get expenses by date range
      parameters:
      - description: Expense date start
        in: formData
        name: date_start
        required: true
        type: string
      - description: Expense date end
        in: formData
        name: date_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Expense'
            type: array
        "400":
          description: Incorrect parameter
          schema:
            type: objcet
        "404":
          description: No rows
          schema:
            type: objcet
        "500":
          description: Internal server error
          schema:
            type: objcet
      summary: Get expenses by date range
      tags:
      - expense
  /expense/id/{id}:
    delete:
      description: Delete expense by expense_id
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Delete expense by expense_id
      tags:
      - expense
    get:
      description: Get expense by expense_id
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/types.Expense'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get expense by expense_id
      tags:
      - expense
    patch:
      description: Patch expense by expense_id
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Date 'yyyy-mm-dd hh:mm:ss'
        in: formData
        name: expense_date
        type: string
      - description: Amount
        in: formData
        name: expense_amount
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Patch expense
      tags:
      - expense
  /expense/new:
    post:
      description: Create new expense
      parameters:
      - description: Expense date
        in: formData
        name: expense_date
        type: string
      - description: Expense amount
        in: formData
        name: expense_amount
        required: true
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: New expense
          schema:
            $ref: '#/definitions/types.Expense'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Create new expense
      tags:
      - expense
...
produces:
- application/json
swagger: "2.0"